Validation and Exception
Validation of attributes and parameters as well as custom exception implemented

+++ challenge-todolist/src/main/java/br/com/lucamatos/challenge_todolist/controller/TodoController.java
@@ -1,7 +1,9 @@
 package br.com.lucamatos.challenge_todolist.controller;
 
 import br.com.lucamatos.challenge_todolist.entity.Todo;
+import br.com.lucamatos.challenge_todolist.exception.TodoNotFoundException;
 import br.com.lucamatos.challenge_todolist.service.TodoService;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -19,8 +21,9 @@ public class TodoController {
         this.todoService = todoService;
     }
 
+    @ResponseStatus(HttpStatus.CREATED)
     @PostMapping
-    List<Todo> create(@RequestBody Todo todo) {
+    List<Todo> create(@Valid @RequestBody Todo todo) {
         return todoService.create(todo);
     }
 
@@ -29,26 +32,29 @@ public class TodoController {
         return todoService.list();
     }
 
+    @ResponseStatus(HttpStatus.FOUND)
     @GetMapping("{id}")
     public Optional<Todo> findById(@PathVariable("id") Long id) {
         Optional<Todo> todo = todoService.findTodoById(id);
 
         if (todo.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Todo not found.");
+            throw new TodoNotFoundException();
         }
 
         return todo;
     }
 
-
     @PutMapping
-    List<Todo> update(@RequestBody Todo todo) {
+    List<Todo> update(@Valid @RequestBody Todo todo) {
         return todoService.update(todo);
     }
 
     @DeleteMapping("{id}")
     List<Todo> delete(@PathVariable("id") Long id) {
-
+        Optional<Todo> todo = todoService.findTodoById(id);
+        if (todo.isEmpty()) {
+            throw new TodoNotFoundException();
+        }
         return todoService.delete(id);
     }
